% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pigify}
\alias{pigify}
\title{Translate a text to Pig Latin}
\usage{
pigify(text)
}
\arguments{
\item{text}{A character string to be translated}
}
\value{
A character string
}
\description{
Translate a text to Pig Latin
}
\examples{
pigify("This is a short sentence. And this (the second sentence) is another one.")
}
