% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pigify_word}
\alias{pigify_word}
\title{Translate word to Pig Latin}
\usage{
pigify_word(word)
}
\arguments{
\item{word}{A character string to be translated}
}
\value{
A character string
}
\description{
Translate word to Pig Latin
}
