% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{word_apply}
\alias{word_apply}
\title{Apply function to all words in a sentence, maintaining proper capitalization}
\usage{
word_apply(text, f, ...)
}
\arguments{
\item{text}{Character string}

\item{f}{Function taking and returning a character string}

\item{...}{Arguments passed to f}
}
\value{
A character string
}
\description{
Apply function to all words in a sentence, maintaining proper capitalization
}
